loader animation on an api call is working fine but Spinner Component from the bootstrap is not working
