+++ frontend/src/App.js
@@ -1,24 +1,10 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import './App.css';
-import { allFollowers } from './Api/Api';
+import User from './Components/Users/User';
 const App = () => {
-  const [data, setData] = useState([]);
-  const allFollower = async () => {
-    const data = await allFollowers();
-    setData(data?.data?.data);
-  }
-  useEffect(() => {
-    allFollower();
-  },[])
   return (
     <>
-      {
-        data.map((data, index) => {
-          return (
-            <h1 key={index}>{data.first_name}</h1>
-          );
-        })
-      }
+      <User />
     </>
   );
 }
