+++ frontend/src/Components/Loader/Loader.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import "./Loader.css";
+const Loader = () => {
+    return (
+        <div>
+            <div className="loading">
+                <div className="loading-spinner"></div>
+            </div>
+        </div>
+
+
+    );
+};
+
+export default Loader;
